+++ BasicShapes.h
@@ -15,17 +15,17 @@
 
 class Rect:public shape
 {
-	int hght, wdth;	//height and width of the recangle
+	double hght, wdth;	//height and width of the recangle
 public:
 
-	Rect(game* r_pGame, point ref, int r_hght, int r_wdth);
+	Rect(game* r_pGame, point ref, double r_hght, double r_wdth);
 	virtual void draw() const;
-	void setHeight(int height);
-	void setWidth(int width);
-	int getHeight() const;
-	int getWidth() const;
-	virtual void resizeUp(int factor);
-	virtual void resizeDown(int factor);
+	void setHeight(double height);
+	void setWidth(double width);
+	double getHeight() const;
+	double getWidth() const;
+	virtual void resizeUp(double factor);
+	virtual void resizeDown(double factor);
 	virtual void rotate(double deg);
 };
 
@@ -35,25 +35,25 @@ public:
 class circle :public shape
 {
 	//Add data memebrs for class circle
-	int rad;
+	double rad;
 public:	
-	circle(game* r_pGame, point ref, int r);	//add more parameters for the constructor if needed
+	circle(game* r_pGame, point ref, double r);	//add more parameters for the constructor if needed
 	virtual void draw() const;
-	void setRadius(int radius);
-	int getRadius() const;
-	virtual void resizeUp(int factor);
-	virtual void resizeDown(int factor);
+	void setRadius(double radius);
+	double getRadius() const;
+	virtual void resizeUp(double factor);
+	virtual void resizeDown(double factor);
 	virtual void rotate(double deg);
 };
 
 class triangle :public shape {
-	int base;
+	double base;
 public:
-	triangle(game* r_pGame, point ref, int r_base);
+	triangle(game* r_pGame, point ref, double r_base);
 	virtual void draw() const;
-	void setbase(int BASE);
-	int getbase()const;
-	virtual void resizeUp(int factor);
-	virtual void resizeDown(int factor);
+	void setbase(double BASE);
+	double getbase()const;
+	virtual void resizeUp(double factor);
+	virtual void resizeDown(double factor);
 	virtual void rotate(double deg);
 };
